magit-diff-visit--range-{beginning,end}: new functions
These functions convert the output of `magit-diff--dwim' to something
suitable for `magit-diff-visit' and replace an older ad-hock
implementation that used to be part of that function.

In addition to revision they may also return `staged' and `unstaged',
but `magit-diff-visit' currently does not differentiate these values and
visits the file in the working tree whenever the returned value isn't a
string.  In the future we will visit blobs for un-committed changes too
in some cases and the logic that decides whether to do that will likely
be implemented in the functions added here.

Create `magit-diff-visit--hunk' from code previously located
in `magit-diff-visit-file', making the latter a bit cleaner.